refactor: 버스 출발 정보 섹션 리팩토링
버스 출발 시간을 표시하는 섹션 컴포넌트를 리팩토링한다.
컴포넌트는 섹션 역할을 하지만 아직 세부 정보를 가지고 있어 소스가 약간
복잡하다. 그러므로 버스 출발 부분의 내용들을 별도 컴포넌트로 분리해서
섹션 컴포넌트는 전체적인 레이아웃 구성과 기본 정보만 가지고 있게 만든다.
다음 출발 시간 카드와 그 다음 시간 목록은 별도 컨텐츠 컴포넌트를 만든다.
기존 내용을 섹션 컴포넌트에서 옮긴다. 즐겨찾기 등록, 해제 소스는 다른
소스와 연결할 필요가 없으므로 별도의 훅으로 만들어서 사용한다.
