+++ app/Main/InformationSection/DepartureInfo/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DepartureInfo';
